wikitest: Impove reporter for ./bin/doctest.
doctest utility tests two parts of code:
examples in the docstrings and examples in the documentation.

This commit join the reporter statistic and the failer details in
the end of both testing processes.

Also a few options were enabled for wikitest and doctest utilities
(like for main ./bin/test utility) such as --no-colors, --no-cache,
 --types.