bin/use2to3: Handle extension-less files correctly
The use2to3 tool previously only considered files ending in .py or .txt
to pass to 2to3. However, scripts have no extension and weren't being
passed to 2to3. This also required a hack when fixing the shebang lines.
Handle this by making a list of relevant files with no extension and
checking if the modified file is in it.