wikitest: Enabled doctest's options FUTURE_ONLY. RELEASE_ONLY
With the help of those options we can tune page, so it will be
valid for test proccess if the options are setted correctly.
E.g. for page

    https://github.com/sympy/sympy/wiki/Linear-Algebra-Module

Those commands do not produce fails:

    ./bin/wikitest -f Linear
    ./bin/wikitest -f Linear --against release

At the same time, we can run all tests with the help of '--all'
option. When we test against master version, the command

    ./bin/wikitest -f Linear --all

produce this fail:

    zeros([2,5])        #doctest: +RELEASE_ONLY

and the command:

    ./bin/wikitest -f Linear --against release -all

produce this fail:

    zeros(2,5)          #doctest: +FUTURE_ONLY