wikitest:  's' for the ./bin/doctest and ./bin/wikitest utilities.
For this utilites, count lines wich are marked as +SKIP and so on,
report them as 's' symbol while test process, and count them
for summary.

Note that the standard doctest utility consider a doctests as a few
lines. So the sum of the 'passed' and 'failed' is constant, and
the 'skipped' counter is independed of them.