bin/use2to3: Use git ls-files
Currently, the script traverses the whole source and grabs all .py
files, skipping a few specific directories. If the user has some extra
Python files around they might crash 2to3; also, all files will be
copied to a new directory and this is not optimal. Instead, use
"git ls-files" to get a list of all the files tracked by git and only
copy those. The "old" way is removed.

Fixes issue 2673. Note that this approach is not faster than the old
one.