Fix doctest failures in the tutorial.bg.txt (Python3).
Use init_printing() instead of setting the displayhook.

This failures raise due to the incorrect source of tutorial.txt which was used
for the translation: from release 0.7.1.

But meanwhile, in the development of the script use2to3, the usage of
init_printing() was corrected in /doc/*txt files already.
See commit 0319a9181260e4177c70149dec38a8a3cf6cb98f.