Improve coupled spin algebra to allow arbitrary coupling
This commit modifies existing methods to allow for an arbitrary number
of angular momentum spaces to be treated by the spin algebra. The
primary change in dealing with spin states is the CoupledSpinState
classes now can take lists for jn (previously jvals) of arbitrary length
and they can be passed a 'jcoupling' option, which defines how these
angular momentum spaces couple together. For example:

JzKetCoupled(2, 1, (1,S(1)/2,S(1)/2))

defines a coupled state between three angular momentum spaces with the
default coupling order. To define the order these are coupled, we pass
the jcoupling option, like this:

JzKetCoupled(2, 1, (1,S(1)/2,S(1)/2), jcoupling=((1,3,S(1)/2),) )

This defines the state to be constructed the j1 and j3 spaces to a
space j13=1/2. This coupling is taken into consideration when uncoupling
the term and is defined by the uncouple method when dealing with tensor
product states of more than two states.