Improvements to (un)couple and testing
Forces the final coupling to be explicit when specifying jcoupling,
rather than making the final coupling automatic due to ambiguity and
possible factors of a sign when the automatic coupling methods pick a
different coupling scheme.

Allows expressions to be coupled and uncoupled, allowing linear
combinations of states to be passed to these methods.

Implements coupling tests to check that coupling and uncoupling returns
the same state.

Some minor bug fixes to the coupling and uncoupling algorithm.