Czech transl.: translate the untranslated texts
I used the program poedit to update the translation. Unfortunately, it doesn't
seem to preserve the original formatting, so the diff is quite large. It'd be
nice to have some way to always convert the .po file into some canonical form,
so that only real diffs are committed into git.