Improve handling of qapply'ing symbolic states with basis conversion
This takes care of some previously XFAIL cases of using qapply to
evaluate a spin operator with a symbolic state that requires a sum over
Wigner-D functions to evaluate. This defines a method which deals with
spin operators applied to sums. It also checks that the _apply_op method
is doing something before returning it.