Implement printing tests for quantum module
This commit implements tests for all printing methods in the quantum
module. No changes are made to the printing methods, this is just to
establish a baseline for verifying valid output capable of working in
the ipython Qt console and html notebook.

Note XFAIL tests show where there are errors in the printing methods.