Fix Integer.is_prime
This fixes issue 2964.  Integer.is_prime returned None, except for S.One,
which returned True!  This was fixed to use isprime().  Note that is_composite
still uses integer & positive & !prime, so S(1).is_composite will return True.
I don't think this can be fixed with the new assumptions system, so I just
created an XFAIL test for it for now. Note that this was right before, but
only because S(1).is_prime was wrong.  So I think if one must be wrong,
composite is better (and anyway, this is just a side-effect of using isprime()
in _eval_is_prime).