Very mild optimization, O(1) for O(n)
I don't know how often this function is called, but this O(1) lookup is
faster even in the best case of the set of ifs it replaces.  In the
worst case (rop == 'ge') it's more than 3 times as fast.  In my use case
where I use mainly inequalities (and lots of them), I was able to
observe about a 2% decrease in time in my relatively small model.