Issue 2982: fix galgebra for python 3
This pull removes the use of types from GA, and instead explicitly gives
the typename. This fixes a bug where use2to3 takes types.StringType to
bytes, which works with python 2, but not with python 3. The patch was
generated with the following command:
2to3 -w -n -f types sympy/galgebra/GA.py
where the instances of 'bytes' was then changed to 'str'.

This addresses Issue 2982:
http://code.google.com/p/sympy/issues/detail?id=2982