Creating symbolic nsolve expression.
When the equation in nsolve contains too many free symbols the nsolve is
not evaluated but it's placed in a lambda within an implemented_funtion.
That implemented_function is called on the free symbols and the resulting
expression is returned.

Examples
========
>>> r = nsolve(tanh(y*x)-x, x, 1, imp_function='root_of_equ')
>>> r
root_of_equ(y)
>>> r.evalf(subs={y : 0.9})
0
>>> r.evalf(subs={y : 1.1})
0.502940574944642

Problems
========
evalf() does not play well with sympy.mpmath.matrices.matrices.matrix.
The tests for this are commented out.
An issue was created for the problem: issue 2994

In [14]: r2.evalf(subs={y : 1.2})
Out[14]: root2(y) # not as expected
In [15]: r2._imp_(1.2)
Out[15]: # as expected
[ 1.2]
[-1.2]
In [16]: type(_)
Out[16]: sympy.mpmath.matrices.matrices.matrix

* Default choice of name for the nsolve expression.

When you call nsolve with too many free symbols it will create an
implemented_function with a default name if you do not supply one.

In [1]: nsolve(x-y,x,0)
Out[1]: nsolve_root(y)

In [2]: _.subs({y:1})
Out[2]: nsolve_root(1)

In [3]: _.evalf()
Out[3]: 1.0000000000000

* Workaround for the creation of nsolve symbolic expression from Matrix.

Two problems were present:

- sympify does not support Matrix
- Matrix and list do not have free_symbols

Both of those are casted to Tuples as a workaround:

f = Tuple(*tuple(f)) if isinstance(f, (Matrix, list)) else sympify(f)

* Exchange wrongly used atoms(Symbol) in nsolve with free_symbols.

nsolve permits the user to not specify the variable to solve for.
To find the right variable it's poping one from atoms(Symbol).
This will cause an error if solving for an Integral for examples.

Still nsolving for expressions of that type is imposible because
lambdify does not support them.