[physics.units] Wrong form for the atomic mass.
The atomic mass must have the units of kg.
The fact that it has the same value as the molar mass in appropriate
units is just a nice consequence of the definition. (1/12th of the mass
of an atom of C12 or something like that matching the definition of mol)

This fixes issue 1722

Reported-by: Nils Winter <nils.winter@gmail.com>