Fix tests round 5: Not incorrect, but match output
The previous iterations of these tests were not logically incorrect, but
relied on all Inequality relations being treated uniformly as less than.
Unfortunately, while (x < y) properly compares equal to (y > x), the And
class is not yet as general as it needs to be.  Currently:

>>> And(Gt(x, -1), Lt(x, 1))  !=  And(Lt(-1, x), Lt(x, 1))
True
>>> And(Gt(x, -1), Lt(x, 1)).compare( And(Lt(-1, x), Lt(x, 1)) )
0

Clearly the first is incorrect, and I suspect to do with the
Basic.__eq__ function through almost all == comparisons pass.  ("Almost
all" does not include the new GreaterThan classes which necessarily
override so they can handle the possibility of comparing a GreaterThan
and a LessThan.)

Another gotcha, in terms of printing at least, is that I don't know how
to make the SymPy internals return a consistent ordering for frozenset
items.  That is:

>>> And(Lt(-1, x), Lt(x, 1)

Could be represented as either of:

x < 1 ∧ -1 < x
-1 < x ∧ x < 1