Apparently for Python 3, must override __hash__
The problem: even though the *Than classes subclass from Expr, they
were not inheriting the __hash__ function, and were thus not hashable.
Based on Tom's observation that:

"I believe the reason here is that GreaterThan redefines __eq__, and
python3 wants you to redefine __hash__ as well in this situation (since
not doing so is often wrong)."

this commit specifically sets up those __hash__ functions since the
values are (until proven otherwise or changed elsewhere) are still
correct.