fixed independent and dependent - added pspace_ind
dependent is now more strict about defining dependence. It is possible
for a two expressions to be independent even if they share common symbols

We now check that the densities are unchanged after information is given about
the other expression. This test unfortunately breaks this functionality for
finite random variables.

This exposed a bug in conditional FRVs. They will raise an error on a symbolic
condition rather than die quietly.

Added pspace_independent which just checks for matching symbols. This is used
internally and never where the difference is significant.