Improved checking of spin states
Implements checks when creating new SpinState's and CoupledSpinState's
to ensure the arguments give a valid state. This includes checking
integer/half-integer arguments and j/m relationships for both coupled
and normal states and checking that the coupling is valid for coupled
states.

In addition, changes the pretty printing of coupled spin states to use
subscripted characters.

Fixes Issue 2973