[series] Fix a gruntz problem.
The basic problem is that e.g. sin(x + sqrt(y)).series(y) returns
non-sensical results, because it tries to do a taylor expansion.

The gruntz algorithm uses certain generalised series expansions which are
allowed (and indeed required) to contain fractional exponents. Producing these
for more or less arbitrary functions seems hard.

Instead, we adapt the `rewrite' function to try and avoid creating fractional
powers. Since this function is free to choose the function in terms of which it
rewrites the expression, this is always possible (provided no irrational
exponents occur). This effectively resolves the problem, since the
above-mentioned problematic series expansions no longer occur.

It is necessary to adapt the function calculate_series, since in certain
pathological cases avoiding the fractional exponents can mean that expanding
to 8th order is not sufficient.

There does not seem to be a satisfactory way to test these changes. The
behaviour of `rewrite` depends on the ordering of the dummys in a dict
it is passed.