docs/outreach.txt: Update links, delete cruft
The outreach.txt file (linked directly from index.txt) is outdated.
Update the links to the SymPy Papers page and our logos, delete the
section about GSoC (as it's covered already in aboutus.txt) and delete
the list of blogs/magazines mentioning SymPy in favor of just a link to
the wiki.