core/relational: Fix Sphinx errors in docstring of GreaterThan
There were several errors in the docstring of GreaterThan; these were
then reported multiple times because the same docstring is used in all
*Than classes. Mostly they were related to using a star (which denotes
the start of emphasized text) without closing it. There were also two
errors about the misuse of references.

Note that the docstring still doesn't render perfectly in Sphinx, but
it's readable and no warnings and errors are reported by Sphinx.