solvers/ode: Fix warnings thrown by Sphinx
The docstring of classify_ode has a non-standard heading, which doesn't
play nice with the numpydoc extension we use for Sphinx. This causes the
whole section not to display in the generated docs. Fix by renaming,
from "Notes on Hint Names" to just "Notes". This renders a workaround in
docscrape.py obsolete, so delete it.

Also correctly indent the docstring of ode_1st_homogeneous_coeff_best to
get rid of another warning.