utilities/autowrap: Remove :ref: tag from docstring
The :ref: tag was used in some Sphinx docs. At some point in the past,
the documentation was moved directly to the .py file, where it is
(apparently) impossible to use the :ref: tag. Remove it and instead just
emphasize the function name. This fixes an error thrown by Sphinx.