Replace all occurances of Sympy with SymPy
The name of the program is SymPy (with a capitalized P), let us try to
be more consistant in using it. This was done with the following script:

if [ -z "$3" ]
then
  echo "== Usage:    gsed search_string replace_string [path]"
else
  egrep --exclude-dir=.git -lRZ "$1" $3 | xargs -0 -l sed -i -e
"s/$1/$2/g"
fi

Replacing sympy with SymPy couldn't be done automatically (due to all
the "from sympy import ..." statements in docstrings), so it's not
really worth it.