doc/ext/docscrape: Remove support for non-standard headings
The Numpydoc standard defines some standard headings which should be
used in all docstrings, but we used some which didn't conform to this
standard and this caused errors in Sphinx. The initial approach was to
explicitly add the non-standard headings we used, but this isn't
practical. Instead, remove these from doc/ext/docscrape.py and rename
the offending headings.

The changes were mostly simple - renaming to equivalent, supported
headings (eg. "Yields" to "Returns", "Arguments" to "Parameters") or
using plural ("Note" to "Notes", "Return" to "Returns"). In some cases
the headings were simply dropped and the content folded under the main
summary.