Simplification back on in Kane in mechanics
Testing revealed that it is probably better to have this on, as it will benefit
people who are attempting to model smaller systems. A long warning about the
implications of this is given in the Kane docstring; hopefully the people
trying bigger systems who will be negatively impacted by this change will be
more likely to read the docstring.