A multitude of changes concerning the plotting module.
1.
`plot` now handles lists of expressions, Plot objects and Series objects.

Examples of lists of expressions as arguments:
plot([x,x**2]) # with implicit default range
plot([x,x**2], (0,2)) # explicit range
plot([-abs(x*y)-10, abs(x*y)+10], (-5,5), (-5,5)) # 3D surface

Examples of Plot and Series objects as arguments:
a = plot((x,(0,2)), (x**2,(0,2))) # creating a Plot object
plot(a) # Plotting a Plot objects
plot(a[0]) # Plotting a Series object

2.
Extensive changes to experimental_lambdify.

Now there is a smart vectorized_lambdify function that tries to work around
all the problems of lambdifying.

The following translations are tried:
only numpy
    - on FloatingPointError (numpy tried to return a nan):
        take the real part and use complex arguments
    - on AttributeError (meaning that numpy can not uderstand the object)
      or on TypeError (unhashable type) raised by sympy:
        only python math and then vectorize float64
    - on ValueError (Symbolic value, can't compute):
        only python cmath and then vectorize complex128

When using python math/cmath there is no need for evalf or float/complex
because python math/cmath calls those.

This function never tries to mix numpy directly with evalf because numpy
does not understand sympy Float. If this is needed one can use the
float_wrap_evalf/complex_wrap_evalf options of experimental_lambdify or
better one can be explicit about the dtypes that numpy works with.

3.
Refactor the general options in the matplotlib backend.

The sequence of operations was refactored and simplified. Some of the
operations were redoing previous operations and this was not good.

Also a new auto option for axis_center was created as a consequence.
Now the axis do not go out of the viewing window.

4.
Small fix for np.array reshaping in the plotting module.

Resetting the shape attribute was working in numpy 1.5.1 but was raising
AttributeError: incompatible shape for a non-contiguous array
in numpy 1.6.1.

Now the reshape() method is used instead of changing the shape attribute.

5.
Small cosmetic changes