all Numbers support divmod
When divmod(2, S(3)) is called, Python passes the call off to
Integer.__rdivmod__ because S(3) is an Integer. The arguments
__rdivmod__ receives are rdivmod(S(3), 2) -- note the reversal of
arguments relative to the first call. Since there was no __rdivmod__ for
anything but Integer, any call to divmod with python number first
failed; since there was no __divmod__ written for anything but Integer,
a call to divmod with the first argument anything but an Integer failed.

__rdivmod__ and __divmod__ are now defined in Number so Rationals and
Floats now work, too. The Integer versions of __rdivmod__ and __divmod__
were left in place since they are about 2X faster.

The following lines were deleted from __rdivmod__ because, with the
argument reversal, other is the Integer -- fortunately this always failed
or else the next line would have since self is not a SymPy object (or
else __rdivmod__ wouldn't have been called, self's __divmod__ routine would
have been called.

-        if isinstance(other, Integer):
-            return Tuple(*divmod(other.p, self.p))