(Try to) fix issue 3014
The mathml printer already adds a newline to the end of the output, so the
print statement in print_mathml should have a , at the end.  Based on the
error from that issue, I think this might fix the problem (though I won't know
for sure until Tom tests it, as I cannot reproduce it myself).