Piecewise raises error with bool args and evaluate=False
Change behavior of Piecewise so if bools are given as args and it is
given the keyword evaluate=False, it raises a NotImplementedError. This
can be changed once classes implementing boolean logic that subclass
Basic are implemented. Changes are made in meijerint to parse the args
of a Piecewise function used in that module which may have bool
conditions, but should not be evaluated. This can be changed to use
evaluate=False when the Basic boolean classes are implemented.