Fix for Issue 1740; int_tested now has a strict flag
Instead of asserting n > 0 we check whether it is a symbol and n > 0.

int_tested can be used to check that the input value satisfies the
condition, int(input) == input, when strict=True. If strict is
False then the input is converted to an int, if possible, else an
error is raised.