remove inf recursion possibility
Im Mul._eval_expand_mul, the fraction of the expression is calculated
and if denom is True then d is processed; but upon re-entry, if the
fraction of that gives the inverse of itself, an infinite recursion
condition will result, e.g.

fraction(exp(neg)) -> (1, exp(-neg))
fraction(exp(-neg)) -> (exp(neg), 1)
etc...

So fraction was modified so both positive and negative exponents
are handled when known.