Issue 2855: Move any sorting of arguments from Add.flatten
The code of the method 'flatten` of the class 'sympy.core.add.Add'
 contain the following lines.
    # order args canonically
    # Currently we sort things using hashes, as it is quite fast. A
    # better solution is not to sort things at all - but this needs
    # some more fixing. NOTE: this is used in primitive, too, so if
    # it changes here it should be changed there.

    newseq.sort(key=hash)   <-- this line.

It is related with an Add class constructor (ordering of args before constructing),
 and therefore with basic hash functions which is a hash of tuple
of sorted arguments.

This kind of sorting moved to _hashable_content method of Add class.
So the orders of args of the Add class became as they are created from constructor.