Don't use cycle detection in _what_known_about()
Infinite recursion is now prevented by temporarily setting the value
of the "fact" being evaluated to None, so that recursive calls for
the same fact return None immediately. This value is overwritten if an
actual result is obtained.

The CycleDetected exception and the _a_in_progress attribute can thus
be removed.