Doctest coverage can be run from anywhere
In order to run, the user simply needs to provide correct
paths to the module and the executable. Also it is assumed
that module and script lie in the same hierarchy. We construct
the module name by using os.path.relpath. Also, in the final
reporting -- name.of.module instead of path/to/file is used.