simplify use of simplify flag in matrices
Cleaned up the use of simplify and simplified in matrices.py.

It's confusing to know which word to use to say you want to simplification to be used (and there was one test that had used the wrong keyword). The way I've done it now is to not have two keywords: set simplify to True if you want simplification (defaulting to sympy's simplify) or pass your own function to use that instead. There is no need for two words (simplified to use simplification, defaulting to sympy's simplify and simplify to define a different function). The use of simplified now raises a deprecation warning.

eigenvects now accepts a custom simplifier like the other matrix functions.