Denesting in a class of radicals with depth 3 via biquadratic equation.
Radicals of the form `z = sqrt(a + b*sqrt(r))`
where `a, b, r` are linear combinations of square roots of positive
rationals on the rationals, with `a**2 - b**2*r > 0`,
can be denested in a simple way by solving a biquadratic equation.
Most of the radicals in this class were not denested before by
sqrtdenest.