Fix of issue 2962: implemented correct presentations of D1 and D2.
The dihedral groups Dn for n = 1 and n = 2 are exceptional in the
sense that they cannot be embedded in the symmetric group Sn
as all the other dihedral groups can. This was taken care of
by explicitly providing the natural presentations of the
two groups, as subgroups of S2 and S4 respectively.