3162: Pow.as_real_imag() doesn't return (foo, 0)
It just returns the general result in terms of the re and im portions of
the base. A change is made in _eval_expand_mul to handle the infinite
recursion caused by issue 3022.

The solution was checked numerically to be the same; the former
solution given did not check out numerically.