TableForm: allow per-column alignments
Columns can each be set by sending a list of alignments, e.g. ('c', 'l')

If single letters are being used, these can also be sent as a single
string, e.g. 'cl'

If there are row headings, they will be right justified unless an alignment
is explicitly given for them. So for an n-column table, n+1 formats must
be sent if one wants to control the alignment of the row headings.