[concrete/summations] Implement some binomial sums.
Specifically, in the evaluation method using hypergeometric functions,
when encountering a finite sum from a to b, try to evaluate it by summing
from a to infinity and subtracting the sum from b+1 to infinity.

Fix a problem where _eval_sum_hyper substitutes zero in the expression
and, if it gets zero, adds one and starts over. This yields to an
inifinite loop if the expression is identically zero under the correct
assumptions.

Extend binomial.eval() to recognise that binomial(n, k) is always zero
if k is bigger than n.

Add tests.