Piecewise._eval_interval(): Use an helper routine _sort_expr_cond().
There is an useful part of the code here that sorts the intervals on
the real axis and removes any overlap. We factorize this code out into
an helper function _sort_expr_cond() which can be called from elsewhere.
It is also useful to ask, what is the interval for a given condition?
So this is also implemented here, taking into account the overlaps.