Extend sign to represent the complex signum function.
Don't assume the argument to a signum is a real number. The complex
signum function takes a complex number x and returns the complex number
z = exp(I*a) on the unit circle that is the closest to x.

>>> sign(3)
1
>>> sign(3*I)
I

Also, fix the signum derivative, which used to return zero.

>>> x = Symbol('x', real=True)
>>> sign(x).diff(x)
2*DiracDelta(x)