Refactoring and new tests for the plotting module.
A number of regressions were fixed.

There are now tests and doctest, that check for correct instantialization and
saving to disk of figures. It checks only the `plot()` function.

The `Plot()` class was simplified, as it was trying to be too many things at the
same time.

The `Series()` class was simplified for the same reason and it is now named
`OverloadedSeriesFactory()`.

Most of the docstrings were updated, however there is still much to be wanted
from the documentation of the module.