Correct the docstring of `implemented_function`.
The docstring of `implemented_function` was incorrect wrt what type of
arguments are accepted.

A more explicit warning about the differences between special symbolic
functions implemented by subclassing `Function` and simple one-use numerical
functions created by `implemented_function` was added.

A check was added in order to insure the correctens of the arguments.
A `raises` test was added to insure the validity of the check.