Add arg_order argument to symbolic nsolve to enforce an ordering.
When creating a symbolic nsolve one gets a random ordering for the arguments.
This is generally not a problem as the function object is already called with
the arguments in the correct order and the user is expected to use subs on the
resulting expression.

However if the user wants to use the .func argument there is no easy way to get
the argument order. This commit fixes this by permitting the user to specify
the desired order.