Document the _sympy_ method and use it instead of `converter` in a few places
The sympify function has among others two methods to convert 3rd party objects
to sympy objects. First is the use of the `converter` dict which is useful for
classes upon which we do not have control. The second is the more pythonic
definition of a `_sympy_` method to be used by the `sympify` function.

Beside documenting it better it is now used instead of `converter` for
MatrixBase and for GeometricEntity.