Fix some style idioms in physics.mechanics
There were a few calls of the type `f.__call__(arg)` instead of `f(arg)` and a
few complicated loops over `range()` that were easily converted to much shorter
and readable `reduce` calls and list comprehention.

It also fixed somehow a bug with a recent change in the `__call__` method of
Expr.