3270: explicitly drop x-independent terms; use cancel
In the case of an indefinite integral, the constant portion of the result can
be absorbed into the constant of integration (which sympy does not report).
cancel is now used to make sure that a constant (independent of x) is not
lurking in the expression like that which was identified in the docstring.