Make re and im call as_real_imag() (Fix Issue 754).
Since as_real_imag() uses re and im, its power was limited by the
fact that re and im were not able to call as_real_imag() in their
turn.  To cope with infinite recursion, a new hint was introduced,
which tells as_real_imag() whether to try to invoke re and im on
self or not.