powsimp: don't assume expr type is known; tidy up code
In the Mul-handling of the '# combine is 'base'' block, expand_mul
is applied to expr and it is assumed that the result is either a Mul
or an Add. Tests were added which fail in master because this is not true.

The recurse function was added to tidy up the code and make it clearer
if it is an identical recursion that is happening or if a flag is
changing.