powsimp: use expansion on new_base
This is a hard term to get to in the simplification process. Once the
power is rebuilt and multiplied by something, any terms that were
collected will be autoexpanded again. So the expansion is done here to
see if some terms were joined in the process. This is quantified with a
_terms function that counts how many terms an expression should give
when it is expanded out.

In simplify, the numerator and denominator processing are done unconditionally
now since the expansion is being handled powsimp and more than an Add
can benefit from this.