random-related changes
There were changes to several of the random routines in 3.2 which make
testing results differ between versions of python. Although random() will
always return the same value for the same seed, uses of randint and randrange,
for example, will give different results because of how the number is
generated changed. It was decided to not try backport changes so tests would
always give the same results for any version of python that sympy supports.

There were some other changes that were made in the process of attempting
to make that backport, however:

Used randrange (which does not yield the max value) in perm_groups
since the unrank algorithm is expecting the rank to be 0 <= r < order. randint
*includes* the upper limit given to it, e.g. randint(1,3) will generate 1, 2
or 3 whereas randrange(1,3) will only generate 1 or 2.

The docstring of cycle_length was modified to reference parameter x0
and to site the correct value being used in the doctest