allow sets to be returned from solve
>>> solve([x**2 - 3, y - 1], set=True)
([x, y], set([(sqrt(3), 1), (-sqrt(3), 1)]))

If there is no solution then ([], set()) is returned,
so if you don't care about the symbols you can always
select element 1. There is no guarantee that the symbols
returned will be in the same order that they were input,
so if there are multiple symbols solved for you should
pay attention to the returned list of symbols.