geometry: perpendicular segment and orthocenter corrections
The perpendicular segment may not terminate on the LinearEntity so
the intersection must be done with the line containing the linear
entity. The same sort of reasoning applies to the orthocenter calculation:
since the altitudes are segments, they may not intersect within the triangle
so lines containing the altitudes should be used to find the intersection.