In Morphism, define __new__ to allow keyword arguments.
Since Morphism derives from Basic, and Basic.__new__ does not accept
keyword arguments, Morphism could not be constructed with keyword
arguments.  While this is all right for simple constructors like those
of Object and Category, the constructor of Morphism will quickly become
awkward to use if the user will be required to supply all arguments in
order.