IntConst class for integr. consts and better system of ODEs solver
The "system of ODEs" solver introduced in the last commit is now well
documented and coupled to the `dsolve` function. New functionality was added as
can be seen from the docstring.

There was need for better treatement of the integration constants, thus a new
IntConst class that subclasses Dummy was introduced. There are still many
issues surrounding integration constants, however this is a first step towards
fixing issue 1739.