nsimplify expr: use simultaneous substitution
nsimplify has troubles with the Floats in this expression because while doing
successive substitutions. The expression changes and creates new
Floats that weren't detected in the call to atoms. Using xreplace (with a
dictionary of replacements) does not work either. Simultaneous
substitution solves the problem:

-3.0*z**2*(x**2 + y**2 + z**2)**(-2.5) + 3*(x**2 + y**2 + z**2)**(-1.5)
>>> eq=_
>>> nsimplify(factor(eq.subs(x,0)))
-3.0*z**2/(y**2 + z**2)**(5/2) + 3.0/(y**2 + z**2)**(3/2)

Notice the leading 3.0 which should be a 3. This is fixed by this
commit.