Revert "Compute numerical roots if domain=RR (#1868)"
This reverts commit f638fcfb46e760c22f6a664a1485b7f33fa56247.

Conflicts:

	sympy/polys/polyroots.py
	sympy/polys/tests/test_polyroots.py

The behavior of `roots` was not consistent with what was writtend in the
docstring. `roots` is documented as returning only radicals, however it was
returning also aproximante `float` solutions. `nroots`, `RootOf` and
`Poly.all_roots` already provide the functionality that was added here.

The behavior was changed to the old one, consistent with the
docstring and the docstring itself was updated to point out the standart ways
to get non-radical roots.